Extract history list query, aggregation, and formatting into shared modules
Moves the sanding-summary MQL query, per-day aggregation, status/zone
formatting, and duration formatting out of React components into plain modules
so the same logic can back a non-React consumer (the forthcoming CLI) instead
of being duplicated.
