+++ src/lib/__tests__/passUtils.test.ts
@@ -1,5 +1,10 @@
 import { describe, it, expect } from 'vitest'
-import { getExecutionTimeMs, getStepVideos } from '../passUtils'
+import {
+  formatSelectedZones,
+  getExecutionTimeMs,
+  getPassStatusLabel,
+  getStepVideos,
+} from '../passUtils'
 import type { Pass, Step } from '../types'
 
 const makeStep = (
@@ -34,6 +39,47 @@ describe('getExecutionTimeMs', () => {
 
   it('returns 0 for no steps', () => {
     expect(getExecutionTimeMs(makePass([]))).toBe(0)
+  })
+})
+
+describe('formatSelectedZones', () => {
+  it('strips the zone_ prefix and sorts numerically', () => {
+    // String sort would give "1, 10, 2"; numeric sort is the contract.
+    expect(formatSelectedZones(['zone_10', 'zone_2', 'zone_1'])).toBe('1, 2, 10')
+  })
+
+  it('accepts a single non-array value', () => {
+    expect(formatSelectedZones('zone_3')).toBe('3')
+  })
+
+  it('returns dash when undefined, empty, or non-numeric', () => {
+    expect(formatSelectedZones(undefined)).toBe('—')
+    expect(formatSelectedZones([])).toBe('—')
+    expect(formatSelectedZones(['zone_x', 'nope'])).toBe('—')
+  })
+})
+
+describe('getPassStatusLabel', () => {
+  const withState = (current_state?: string, success = true): Pass => ({
+    ...makePass([]),
+    success,
+    current_state,
+  })
+
+  it('maps terminal current_state values to labels', () => {
+    expect(getPassStatusLabel(withState('Succeeded'), false)).toBe('Success')
+    expect(getPassStatusLabel(withState('Failed'), false)).toBe('Failed')
+    expect(getPassStatusLabel(withState('Cancelled'), false)).toBe('Cancelled')
+  })
+
+  it('distinguishes incomplete from actively running for non-terminal states', () => {
+    expect(getPassStatusLabel(withState('Executing'), true)).toBe('Incomplete')
+    expect(getPassStatusLabel(withState('Executing'), false)).toBe('In Progress')
+  })
+
+  it('falls back to success flag when current_state is absent', () => {
+    expect(getPassStatusLabel(withState(undefined, true), false)).toBe('Success')
+    expect(getPassStatusLabel(withState(undefined, false), false)).toBe('Failed')
   })
 })
